cwa/tests: split tests in main into suite-specific modules
