+++ cwa/tests/src/http.rs
@@ -1,10 +1,13 @@
 extern crate libcwa;
 extern crate httparse;
 
-use libcwa::*;
+use self::libcwa::*;
+use self::libcwa::Resource;
 use std::io::{self, Read, Write};
 
-pub fn test() -> Result<(), i32> {
+pub extern "C" fn test() -> Result<(), i32> {
+    log::info("running HTTP tests");
+
     let reqd = "GET /404 HTTP/1.1\r\nHost: printerfacts.herokuapp.com\r\nUser-Agent: Bit-banging it in rust\r\n\r\n";
     let mut headers = [httparse::EMPTY_HEADER; 16];
     let mut req = httparse::Request::new(&mut headers);
